Moved get_urls for detail view into viewableModeladmin
